Fixes issue where invalid DFDL Entities were simply ignored.  We now SDE on invalid DFDL Entities.
Adds basic tests for this.

Implements escapes for percent sign in
entity replacer.

Noticed bug in Delim parser.
Seems like we want to look for local
delimiters first before remotes. Changed
the entry to the correct order.

Changes TDMLRunner so that the replaceDFDLEntities
attribute is by default set to 'false'.

Updates tests.

DFDL-440