Add the ability to link outside tests to the sbt config
Simply add a symlink in daffodil-extra, linking to the project. The sbt
configuration will find this and treat it as if it is normally included
as a subproject.

Note that nothing in the daffodil-extra directory should be committed.
There is a .gitignore entry to prevent accidentally committing anything,
but it's still possible. Do not force commit anything to override the
.gitignore.

DFDL-753