Fixes issue where removeHiddenElements wasn't working properly.
Also makes it so that the JDOM code is used
even on the Scala side of the house that way
we can ensure the JDOM conversion/removal code
is executed/used by the tests.

removeHiddenElements and removeAttributesJDOM now mutate
the Element passed to them.

Adjusted the 'result' code to coincide with this change.

DFDL-762