Require sbt 0.12.2
Some of the sbt plugins we use were not built for sbt < 0.12. If you
try to run an older version of sbt, you just get unhelpful "Unresolved
Dependency" errors. With this change, if you're not using sbt 0.12.2,
then sbt will download it and then run using the correct versoin.

DFDL-772