This is a combination of 4 commits. Fixes for iso-8859-1, and fixes to debugger to improve robustness a little by improving printing of delimited primitives in trace.
Added test that illustrates how to achieve a final optional trailing
element which contains the sequence separator unescaped.

Fixed annoyance in trace when eval .. for root element. It checks for
this now and if .. is the document, it just uses .

Improve SDE about bit length mismatch to say what the lengths are and
fixed negative test with too-precise error message. Message was improved
to add information, but that broke the expected error message text of
the test.

DFDL-797, DFDL-805