binaryNumberRep="binary" support for xs:decimal
This commit adds binary support for xs:decimal.
Ran into a few issues with BigDecimal trying to use scientific notation,
but that should be working now since toPlainString instead of toString.

Renamed CompilerTunableParameters to DaffodilTunableParameters

This was based on Mike's suggestion that this name makes much more sense
than Compiler and since I was already making changes to this file that I
should go ahead and rename it while I was in there.

DFDL-106