Implemented isLastDeclaredRequiredElementOfSequence
This function mainly comes into play when verifying that an element with
maxOccurs="unbounded" and occursCountKind="implicit" is that last
required element of a sequence. Otherwise, and SDE is given.  So far
only the compile time checks have been finishd.  There are a few cases
where an SDE will need to be generated at runtime that will need to be
added, such as cases with choices, patterns, etc.

DFDL-732