Changes to toBriefXML to make printing of parsers more uniform.
Due to logging or some such, some things were no longer evaluated that
were before, and hence, parsers are failing (asking for properties at
runtime, not compile time). Hoisted the offending into the primitives,
not the parsers.

There was one mysterious thing with a comment to the effect of "can't
move this to the primitive or badshit(tm) happens". I moved it and it
seems to work fine now.