Parse character entities in textStandardZeroRep
Before this change, we were doing a straight string comparison of
textStandardZeroRep and the parsed data. This meant that a zero rep of
"%WSP+;" would only match a literal "%WSP+;" rather than matching one or
more whitespace characters.

This patch modifies the zero rep processing to use delimiter scanning
code to create a regex that will match strings that might contain
character entities, and performs regex matching to determine if a zero
rep exists or not.

DFDL-868