Implements the check for alignment ambiguity.
Adds isOptional flag to ParticleMixin.

Adds checkForAlignmentAmbiguity method
to LocalElementBase.

Adds couldBeNext method which contains
a list of possible terms until the
first required one.

Adds additional unit tests for the
couldBeNext method.

Adds a LCM and GCD method.

Adds alignment comparison for an
optional element or variable occurrence array
and what follows it.

Adds an Assert.invariantFailed to
isDeclaredLastInSequence to show
that this is only a valid question
when the object is enclosed by a
sequence.

DFDL-698