Fixes issue causing BF tests to fail.
UnorderedSequences needed to change the
choiceElement to have min/maxOccurs
reflecting the total minimal number
of elements contained within the sequence
as well as the total maximal number
of elements.

Added a checkOccurrence cal to the
UnorderedSequence parser.  This issues
a processing error when the min/maxOccur
requirements aren't met.

Fixes issue where the BE001 test incorrectly
had 5 z elements instead of the max of 4.

This change is necessary as we've added the
min/maxOccurs check to the UnorderedSequence
parser.

DFDL-180