Use a ListMap instead of a Map to maintain import/include schema ordering
A scala Map has no guarantee of order. However, a ListMap uses a
list-based data structure for implementing the map inteface, which
maintains insertion order. By changing the data structure of the IIMap,
order is maintained, and so we should get the correct schema when order
matters (for example, searching for a root node when one is not given).

Also, revert changes in commit ea8b1d8e00eb0bb9ea342542ee904cad26a9ee6f,
which created a temporary workaround.

DFDL-883