Computes the maximum delimiter length of all in-scope terminating markup.
If a delimiter is not found, the max
length value is used to display to
the user what was found instead of
the delimiter.

Factors out delimiter length computation
to a DelimiterText trait that's inherited by
both Static/DynamicText.

Separates maxDelimiterLength computation
into compile-time and run-time values.

Added the maxLengthForVariableLengthDelimiterDisplay
constant to DaffodilTunableParameters.

DFDL-756