Perform and SDE check for complex types in checkConstraints earlier
The call to e.primType right above the SDE check performs a cast to a
SimpleTypeBase. If the context is not a simple type, that cast will fail
with a ClassCastException before it can throw the SDE. So the SDE check
needs to be moved before e.primType is evaluated.

DFDL-921