Performance work suggested introducing the encodingErrorPolicy property, and a new property that is required almost all the time breaks many many tests. That's why so many files were touched.
Also, the definition of daffodilTest1 that the tdml runner was inserting
was different from the one in the built-in-formats.xsd file. They are
now the same - uses an ordinary xs:include now.

Added daffodilNewProperties1 format. This is where new properties should
be added as we introduce them.

In interests of performance, text alignment is broken, i.e., a text
element after a binary bits element that doesn't end on a byte boundary,
will not be aligned automatically. A runtime error occurs that the
characters are not aligned.

Added test for the even trickier version of this, which is the binary
bits fields own terminator (which is textual, even though the element is
bits) isn't going to be properly aligned (this is bug DFDL-929).

Had to modify propgen because it now uses the exact same schemas as are
in daffodil-lib to generate from. Unfortunately there were things that
we added to daffodil-lib's schemas that caused incorrect things to be
generated.

Fixed a few more incorrect daf namespaces.

All tests pass with sbt.
DFDL-648