Scala 2.10 port.
Also included are changes in preparation for bitOrder property needed
for support of MIL-STD-2045-47001D format.

Includes changes to make CR (0xD) character appear as 0xE00D character
in the DFDL infoset.
    
Turns out XML eats CRLFs and replaces them with LF, and converts
isolated CR to LF. So while CR is legal in XML, it's never
part of the XML Infoset.

So, in order to support it as part of the DFDL infoset, we have
to treat it like the xml illegal chars and map it into the Unicode
PUA area. Hence, 0xE00D, not 0xD. 

Also some test changes to better characterize the odd (buggy)
behavior of scala.xml.Utility.trim() that the TDML runner calls.

Turns out this does more than collapse whitespaces. It seems to 
flat out remove them in some cases.

Added test for Scala's trim showing that it is faulty. 

Created our own trim replacement.

Also there's a scala bug in that xml &apos; is not supported.

The test testChars2 illustrates how, in a TDML file, you can 
represent all 256 bytes of iso-8859-1 in data, and how you can
represent those same characters in a tdml expected infoset value.

Converted a large number of tdml and xsd files to use Unix line endings
not Windows CRLF style. Verified that there are no regressions due to 
this change. This eliminates line-ending sensitivities where the tdml
file contained data inside CDATA regions which happened to have Windows
line-endings, in that case the CRs seem to be preserved into the data,
and as a result tests were wanting to see 0xE00D in the result.

This is too fragile - line-endings are usually invisible. So having a
tdml file depend on being saved in Windows CRLF style in order for the
tests to work is problematic. 

Also: Computation of termsUntilFirstRequiredTerm had bug.

Also: update gitignore to ignore eclipse-projects/*/lib directories
(which are supposed to be eclipse symlinks to the real project lib
directory.)

DFDL-941, DFDL-938, DFDL-896