Remove duplicate DFDL_partX schemas
Two copies of the DFDL_partX schemas have existed in both
daffodil-propgen and daffodil-lib. Becuase of this separation,
differences have begun to emerge when we forgot to make changes in both
places. Now that those differences have been fixed by commit
3570eff217c5709c74f83e4ff8ee716e11589b9b, it is a good time to get rid
of the duplicates.

This commit removes the schemas from daffodil-lib and adds a
'gen-schemas' sbt command (similar to 'gen-props') which copies the schemas
from the daffodil-propgen project so daffodil-lib can use them. Note
that similar to GeneratedCode.scala, the schemas copied to daffodil-lib
should never be modified. They should only be changed in the propgen
code, and copied to daffodil-lib with gen-schemas.

Additionally, 'gen-managed' is added as a convenience command, which
will run both gen-props and gen-schemas.

DFDL-899