Simplification of TextDelimitedParser. Removes unnecessary use of Either.
Corrects EscapeBlock delimited parser to parse
as if the escape scheme isn't there if a start block
is not found.

If a start block is found and an end block is not found
the parse fails.

Contains commented out code in an attempt
to move the padding parsing into its own
primitive.  May be useful after Mike's
refactoring of the primitives.

DFDL-175