This commit fixes a couple of different tickets:
DFDL-714:  There were problems importing the dfdl schema when it was split into three separate files.  As a workaround, we per-merged them into a single schema file, but this is not the ideal solution.  This commit fixes the problem in how we were resolving schema namespaces so that we can now support the dfdl schema split into three parts.

DFDL-969:  Simialarly, we had issues resolving schemas that were imported into another schema.  The result was an error stating that the application could not locate a particular schema file.

In addition to the code changes, this commit also contains updates to the dfdl schema file to bring them up to date with the current version as defined in the DFDL_model_all_parts.xsd file.
There is other fixex to a path specified in the tresys-contributed tdml files and the TestXMLCatalogAndValidate test that are necessary as a result of the changes to the schema resolution code.

I had to move some tests to debug because they were importing a schema that was invalid and would not parse.  This was causes some of the tests to fail for the wrong reason, and others to fail that should have passed.

DFDL-714
DFDL-969