Switch to a different JUnit XML reporter plugin
This is the plugin that generates XML reports for our tests that Bamboo
can understand. The old plugin has stopped created reports correctly,
causing it to look like we have far fewer tests than we actually do.
Plus, development of the plugin does not seem all that active. Switch
to a different plugin, which works and has a much more active community
surrounding it.