Move thread local state into a mutable pstate class
The thread local code recently added doesn't really need to be thread
local. We really just want to avoid copying all the data that was made
thread local when we copy PState. So, rather than using thread local,
which has some overhead, add a mutable Pstate (MPState) that stores
information that is mutable, and store a pointer to that in PState. This
limits a copy of just this single MPState pointer while avoiding the
potential overhead of thread local.

Eventually, we maybe be able to entirely switch to this mutable PState
and completely avoid copying.

DFDL-971