Maybe class
Removed ability to pattern match against it. Because this boxes the
objects, and the whole point of this class is NOT to box objects. This
is the most important change - because we may start using this class
heavily, and I don't want us to be writing match-cases when we do. 

Improved inlining slightly. Changed perf test (commented out) to use
Long, not Int. (Makes no real difference in relative speed.).

Fixed 'bug' where def empty was allocating multiple different Nope
objects instead of their only being a single one. 

Confirmed (on JVisualVM) that one can pass and return Maybe objects
without allocation occuring.

Added implicit conversion to an Iterable/List. (Convenience - scala
Option has this.)

Added implicit conversion from a Scala Option to a Maybe (but the other
way must be explicit)