Improve compatability with IBM DFDL
Daffodil supports (and current requires) the new bitOrder and
encodingErrorPolicy properties, which are currently just an errata.
However, IBM DFDL does not yet support these new properties. This means
schemas that work with IBM DFDL but are missing the new properties will
fail if parsed with Daffodil. In order to support older schemas and
maintain compatability with IBM DFDL, make the bitOrder and
encodingErrorPolicy properties optional, and provide a default value if
not given. The bitOrder property defaults to 'mostSignificantBitFirst',
and encodingErrorPolicy defaults to 'replace'.

This also adds the ability to configure tunable parameters.

New Java/Scala API functions are added to the Compiler, which take
either a single tunable name/value pair, or it takes a map of tunable
name/value pairs.

  addTunables(Map[String, String]): Unit
  addTunable(String, String): Unit

A new -T option is added to the command line, which allows one to specify a
key/value pair, where the key is a tunable name and the value is the
value to set it to, for example:

  -TmaxOccursBounds=1024

Also adds new syntax to the Daffodil configuration file to specify
tunables. For example:

  <daf:tunables>
    <daf:maxOccursBounds>1024</daf:maxOccursBounds>
  </daf:tunables>

If both -c (a configuration file) and -T are provided, key/values from
-T overrule key/values from -c.

Finally, add new tunables (requireBitOrderProperty,
requiredEncodingErrorPolicyProperty) to configure whether or not
daffodil should be strict about requiring the bitOrder and
encodingErrorPolicy properties. These both default to a lax behavior.

DFDL-994