Fix version detction with ambiguous refs
The command 'git rev-parse --abbre-ref HEAD', which is used to get the
name of the current branch, will give inconsistent results if there are
tags and branches with the same name (which we have in our workflow). So
instead, use --symbolic-full-name which will always return an
unambiguous name of a branch so we can get the determine the correct
version.

DFDL-1022