Switch from daffodil-pack to sbt native packaging plugin
The daffodil-pack subproject (which isn't really a subproject) is a pain
to update, has code duplicated in build.scala, and really just isn't the
right solution. Instead we should be using a plugin that does
everything. Fortunately, the sbt native packaging plugin is now pretty
mature and does everything we need, and more. So replace daffodil-pack
with the plugin.

This also has the added benefit in that CLI tests now use the exact same
script files that are part of the zip/tar releases. This means things
like setting DAFFODIL_CLASSPATH can now be used in CLI tests. Note that
running sbt stage will now store the script in:

  daffodil-cli/target/universal/stage/daffodil(.bat)?

The CLI tests have been updated to use this path and to append the .bat
when run on Windows.

This also now uses different commands to pack. They have changed to:

  sbt universal:packageBin          # creates .zip
  sbt universal:packageZipTarball   # creates .tgz

Note that we release tar.bz, but the conversion can be easily performed
by running:

  gunzip --stdout daffodil-VERSION.tgz | bzip2 > daffodil-VERSION.tar.gz2

The sbt configuration is also changed to move managed library jars from
lib/ to lib/jars/ and moves managed source jars to lib/srcs/. This way,
jars needed as dependencies are included in the release zips, and those
that aren't (like sources) are omitted.

Note that this plugin also supports creating installers for various
distributions (e.g. rpm, deb, MSI), so we could eventually release those
as well.

DFDL-1033