Centralized SimpleTypeNode and NodeInfo to the NodeInfo file. Renamed SimpleTypeNode to just TypeNode which is more correct.
Changed PrimType objects so they are NOT SchemaComponents. This was what
drove the split of PrimType and RuntimePrimType, so those can be
collapsed back together... next.

Collapsed PrimType and PrimitiveType, and they are no-longer
SchemaComponents so also collapsed RuntimePrimType.

Consolidated PrimType and NodeInfo objects/types/enums. There is now
exactly one file, NodeInfo.scala, that defines types and their
relationships.