Adds support for fn:empty.
Added NodeInfo.Exists. This is now the
targetType for both the fn:empty and fn:exists
functions.

Conversions needed to be altered to include
a case where the sourceType (st) could be
anything but the targetType (tt) was Exists.
In these cases we do not care about the st
as we're only concerned if it exists or not.

Added DownArrayExists to determine when an
array is defined.  If it is defined and
has at least a length of 1 we are good.

Moves working tests from debug to new.

DFDL-1083