Update to latest sbt-pgp plugin
It looks like dependencies with bouncy castle changed, and now the
sbt-pgp plugin, used for signing published releases, is broken. So
update to the latest sbt-pgp plugin. Also, change configurations so we
can just run 'sbt publish-signed' and it will only sign/publish
artifacts that don't have the 'noub' settings. Also configure it to not
look in ~/.gnupg and to always use keys in ~/.sbt.

DFDL-1130