Implements fn:local-name.
Corrects a test of fn:local-name to not use
parent::child notation.  This is not yet
supported and as we are testing fn:local-name
this was changed to be a path.

Moved working tests from Debug to New.

There is one test local_name07 that fails
due to a slotIndexInParent error.  Left this
test in Debug as it fails for a problem
other than that illustrated in DFDL-1096.

Adds check and error message for when local-name
is not a valid NCName.

DFDL-1096