XPath spec states that fn:not uses fn:boolean to turn the arg into an effective boolean value.
The behavior of fn:boolean is different than that
of xs:boolean.  This corrects fn:not to use the
fn:boolean conversion functionality rather than
that of xs:boolean.

Adds an entry to the conversions code to allow for
the 'conversion' of Any type to AnyAtomic.

Basically, we should be able to pass any type
as an AnyAtomic.

DFDL-1074