Fix and/or expression to allow repetitions
Expressions like this are valid:

  { expr1 or expr2 or expr3 }

In these kinds of and/or expressions, the three sub-expressions were all
passed to the OrExpression class. However, the OrExpression class is a
BinaryExpMixin, which expects there to only be two sub-expressions. This
led to the third expression to be completely ignored. This patch changes
it so when there are more than one sub-expressions to and/or operators,
the parser matches are folded so expression essentially changes to

  { (expr1 or expr2) or expr3 }

so that there are only ever two expressions provided to the binary
operator. This also adds an assertion to ensure we always pass two and
only two expressions to a BinaryExpMixin and don't hit this bug in the
future.

DFDL-1182