Make DPath logic expression use short-circuit evaluation
From the XPath spec, "The order in which the operands of a logical
expression are evaluated is implementation dependent."

Arguably, more people are familiar with and expect short-circuit
evaluation. Additionlly, it really only matters when there are
side-effects, which is only possible in DFDL with things like processing
errors. This isn't the most common case, and is still possible to
achieve. Additionally, short-circuit evaluation can also result in
faster expression evaluation since it can bypass some evaluations in
some cases. Because of this, use short-circuit evaluation for Daffodil
logical expression evaluation.

DFDL-1181