Unzip HTML version of DFDL specification
Two reasons:

1) Git handles text much better than binary. This could potentially
allow seeing how the spec changes between versions using just git diff,
which could be helpful

2) The continuous integration Bamboo servers on NCSA have a method to
include artifacts and have them appear online as a "Manual". But we need
the uncompressed HTML and images to be marked as an artifact. Instead of
uncompressing as part of the build process for every build, just leave
it always uncompressed, and potentially get the benefits from 1.

DFDL-911