JSON5 schema and tests.
This is a bug reported by James Garriss of Mitre. There's something
wrong with delimiting by %WSP*;,%WSP*; (i.e., a comma, with optional
whitespace on either side). This is a very reasonable and expected
idiom.

This is a 5-level deep (max) schema for JSON-like data. It is nested
such that there are 560 elements in the schema (not tiny), but it
compiles much too slowly for something that size.

So there are two problems: it fails to handle even tiny examples of json
data - a few such tests are in the tdml, and it is so slow to compile
it's a bit absurd.

DFDL-1196 is the bug report. This is not fixed here. I'm just pushing
the DFDL schema for this and associated tests.