Fix non thread safe code in text number converters
In the static number format factory, we generated a single NumberFormat
and then used that for all thread local numFormat values. However, this
meant that even though we were using thread locals, the values returned
from the thread locals were all exactly the same, which caused threading
issues. This change moves the generateNumFormat code into the thread
local initialValue() function so that a new number format is created for
each thread, which fixes the threading issues.

Also change generatedNumFormat and numFormat in the dynamic format
factor to just be vals. They are variables inside a function and so they
do not need to be lazy or transient.

DFDL-1209