Return 0 or 1 for performance CLI error code
On POSIX systems, the return code is only 8 bits, and so can only be in
the range of 0 to 255. We currently return the number of tests that
failed as the error code, which could easily be above 255. This just
gets truncated so it is possible for the error code to be zero even if
there are errors. So instead, just return an error code of 1 if there
are any errors.

DFDL-1208