Properly set environment variables in CLI tests
The previous code completely overwrote environment variables when
executing CLI tests. This meant that variables like JAVA_HOME and PATh
were completely lost. The build server relies on these environment
variables (e.g. java is not in a standard location), which led to CLI
tests failing because java could not be found.

This change makes it so CLI tests inherit the existing environment
variables, overriding only those that are specified by the test. This
requires a change of the envp from an Array to a Map, and then
converting to an Array for the exec method.

DFDL-1203