Contains fixes for JSON5 user test failures.
Issue appears to be two-fold:

Initiator needs to always be scanned for regardless
of whether or not 'foundDelimiter' is populated. This
is because the 'initiator' is never part of the
terminating markup.

DFAField's DelimsMatcher was using the 'checkMatch'
method for WSPStar.  This was not enough as it would
cause the 'couldBeFirstChar' method to fail if a
space was not matched.  In the case when WSPStar is
the first state, this would always cause the delimiter
match to fail.  Instead in the case when a space was
not matched, we need to ask the next state if this
character matches. If so, the character could be the
start of this delimiter.  If not, then it's not.

Corrects schema to have Null element contain
'nillable' attribute.  Otherwise LiteralNil parsers
were not being used/generated.

Corrects TDML expected test results.

Moves working JSON5 tests from debug to new.

DFDL-1196