Adds ability to specify temp directory for schema files.
Updates createTempFile call in XMLUtils to take a File
representing the temp directory.  If this value is 'null'
it will result to the default directory as specified in
the system property java.io.tmpdir

Makes other necessary changes to other files to pass in
an Option[File] for tmpDir.

At this point, the temporary files are only written to
disk for the TDMLRunner tests.

Adds a test for specifying the tmpDir to which schemas are written.

Test specifies the tmp dir.
Specifies the tmp dir be deleted on exit.
Executes the compiler which then writes to the directory.
Verifies that a file was written to the directory.

DFDL-1103