Improve performance related to ProcessingErrors
ProcessingErrors extend Exception because we sometimes use them for flow
control (e.g. withParseErrorThrowing). Unfortunately, Exceptions are
fairly expensive to use, but this is mostly due to the need to create
stack traces. Since we only use ProcessingErrors for flow control, and
don't care about the actual stack traces, we can override the
fillInStackTrace function to do nothing, and greatly reduce the expense
of creating ProcessingErrors, which leads to a big performance gain.

Create a ThinThrowable trait which can be mixed in to any Throwables for
easy reuse in other places.

Gains with initial testing of files that heavily used assertions and
backtracking, such as PCAP, saw performance gains of around 20%.

DFDL-1224