Change properties Enum to use a Map instead of a List
Change the _values list to a map so that we can get constant time lookup
of strings to Enum values. Additionally, override the toString method so
that it stores the toString value, instead of performing an expensive
lookup each time.

Results in about a 5% performance increase for files that convert
runtime values to enum values, e.g. runtime byte order in PCAP files.

DFDL-1225