Removed extra closing paren in calendarTimeZone regex. This error was spotted in the spec recently, and appears also in our schemas for DFDL annotations.
Fixing this doesn't appear to have any effect on anything. The code
generator just treats this as a string. So this regex is only used when
validating a Long-form DFDL annotation using xerces. 

It seems we have no long-form annotations that have this property. That
or the error in the regex is just tolerated.