Removed split of grammar at unparserDocumentElement and parserDocumentElement, which may improve schema compile speed, but also is just cleaner.
Had to add a dynamic variable though which is used so that the
input/outputValueCalc grammar rules can do the right thing depending on
whether we are getting a parser or unparser.

This doesn't fix DFDL-1238, but contributes to fixing it.

Also fixed logging calls that were evaluating their arguments (causing
giant toStrings to be produced even when not logging)

Also hack-fixed DFDL-1241 - toBriefXML crashing JVM when string is too
big for it. This is a hack. (Separate DFDL-1013 is about getting rid of
toBriefXML entirely, replacing it by a real toXML and a pretty printer
that elides at some depth.)


DFDL-1238, DFDL-1242, DFDL-1241