CLI now pretty prints small infosets, but directly writes out larger infosets without converting to XML first.
DFDL-1232 performance problem - was in converting and pretty printing
the XML. Was many minutes. Now 10 seconds on my laptop, that is, if you
direct the output to a file. (The XML string output is 15+ MBytes)

DFDL-1125 (excess namespace prefixes in infoset) - this makes everything
we convert to XML smaller, and eliminates one of the reasons for a full
tree-walk of the infoset (to remove the namespace bindings). Now the
schema compiler computes the minimum namespace bindings needed for each
element when converted to XML, and this minimized binding set is carried
on the ElementRuntimeData and used whenever we synthesize XML nodes from
the Infoset. Also, it now only puts namespace bindings in place that the
element itself or its children could need, regardless of how many exist
on the DFDL schema XML.

DFDL-1232, DFDL-1125