Removed all calls to Utility.trim in TDML runner.
Added explicit code to strip out entirely-whitespace nodes from expected
results in TDML runner.

We had a number of tests that were written in a way that was depending
on the TDML runner trimming the expected and actual results. Now that
that trimming isn't occurring, tests had to be adapted to the new
(correct) behavior. This allows tests to properly insist on specific
leading or trailing whitespace, but also requires tests to be explicit
about the leading or trailing expected whitespace.

Review comments incorporated.