Fix performance regression with calendar types
Before this change, the infoset string representation of all calendar
types was always calculated and stored in our DFDL calendar
representation. Additionally, when it perform this calculate, it created
a new SimpleDateFormat. It is not always necessary to create the infoset
representation, and additionally, it is expensive to create a
SimpleDateFormat. Since there are only a handful of SimpleDateFormats
for the infoset (e.g. date, time, dateTime, with and without timezones),
make those constants, using thread locals to ensure there are no race
condition. Additionally, no longer store the string representation,
and instead generate when required when the toString method is called.

This resulted in date/time performance on par with daffodil 0.15.0, and
looks like may even be 2-3% faster.

DFDL-1239