Fix TDML runner - detects and warns about duplicate test names now.
Also, put in noisy print statement every time a test suite object is
created. Turns out we are creating (which means re-reading and
validating, etc.) the entire TDML file for every single test.

The fix to this is to reorganize the way we create the scala test suite
files to create the test suite objects once. See TestFacets.scala for an
example of what we need to do.

I've fixed a couple of the large test files. The others still have to be
fixed. This speeds up testing a lot so it is worth the headache.