Corrected tests that needed fillByte or textPadKind now because of padding for fixed length strings.
Changed built-in-formats fillByte to be a space character, not NUL. This
is more suitable for text formats, and the basic format is text.

Added Entity replacement so fillByte of "%SP;" or "%NUL;" will work, in
addition to "%#rXX;" byte value entitiy.

Also improved code base....Used to be that there were subtle circularity
bugs between runtimeData and encodingInfo. I.e., if you used one of them
in the wrong place you'd get an OOLAG circularity detect.

This is fixed now. The encodingInfo object is now created by the
daffodil-core schema compiler, and it is part of the TermRuntimeData
class.

This means we no longer must pass around both a runtimeData object and a
separate encodingInfo object. You can get the encoding info object off
of any TermRuntimeData. That is, off an ElementRuntimeData or a
ModelGroupRuntimeData object.

However, I did not change everything so that the explicity passing of
encodingInfo goes away. Only for the SchemaSetRuntimeData object.

Added ability to inquire whether a unparse result is 'scannable' meaning
is all text in a uniform (and known) encoding. In that case the TDML
runner treats it as text and the diagnostic messages will be textual.

Otherwise it uses hex, (not as previously decimal) values.

Improved diagnostic message part about the schema file locations. No
longer shows One(...) and Nope, and List(....).

Cleaned up separation of EncodingInfo (now named EncodingRuntimeData for
clarity), and TermEncodingMixin, which is the schema compiler's view of
the same things.  Shared things are in KnownEncodingMixin which is part
of the runtime.

Got rid of treating HexBinary as a kind of String. (Mostly)

This was needed because we were demanding text-oriented properties to
exist even for hexBinary data. (E.g., textPadKind, textTrimKind -
neither of which is supposed to be needed for hexBinary).

This is still done for delimited binary (which is allowed), which for
parsing, treats the data as iso-8859-1.

But this idiom will break down for unparsing since we don't want padding
with characters for hexBinary, we want filling with fillByte. But the
string logic for this won't work right, because it only fills in what it
can't pad. It can't pad hexBinary has no padding character (I suppose we
could fake it by making the padChar be the same character as the value
of the fillByte....)

Got rid of duplicated code for removing padding across several traits
and classes.

Got rid of imperative code that computed padChar as a string - was
causing unecessary dependencies on text properties to be created.

Standardized names and conventions for pad characters to parsingPadChar
and unparsingPadChar

We need both, because in the same schema you can have
textPadKind="padChar" but textTrimKind="none", so there can't be just
one pad char object if it is to carry information about both whether or
not a pad character is to be used, and the value.

Padded is a trait used at schema compile time.

Renamed HasPadding to PaddingRuntimeMixin. It is a runtime trait used by
parsers (perhaps unparsers in the future)
Added test that will break if iso-8859-1 behavior changes from what
we're depending on which is that it is a 256 character set, and each
byte maps to the corresponding unicode codepoint directly.

DFDL-1293, DFDL-1294, DFDL-1306