Fix reading unparser infoset from standard in
When reading from standard input, the CLI created a
InputStreamSchemaSource. This had a bug in which is was not providing a
FileInputStream when creating an InputSource, but was instead just
passing in the URI string as the systemID. However, we then changed the
symtemId to "standardInput". So it then tried to resolve to a file
called "standardInput" relative to the CWD. That didn't exist, so it
failed. The solution is to just create a FileInputStream for the file
and passing that into the new InputSource.

DFDL-1216