Ensure infoset output via toWriter() matches toXML()
Two differences were fixed:

1) The root element was duplicated because the DIDocument element was
written (which has the same erd as the root element) as well as the root
element. Instead, for the DIDocument, override the toWriter method to do
nothing and just call toWriter on the root. This mimics the toXML method
for DIDocument, and prevents a duplicate root element.

2) The XML namespaces were always printed for every element. This was
unnecessary and is changed to only print the namespaces that are unique
to the element.

Additionally, a tunable was added, prettyPrintElementLimit, which can be
used to set the infoset size at which to disable pretty printing when
outputting the infoset. This makes testing the above changes easier, by
just setting th limit to zero and thus disabling pretty printing.

DFDL-1322, DFDL-1323