Fix backtrack bug.
Found one of the repetition parsers wasn't duplicating state for use
when backtracking. Once that was fixed, the copying of state in
InStream.scala can be removed.

This fixed occursCountKind='parsed' which was requiring minOccurs
occurrences - which is wrong. occursCountKind='implicit' requires that,
but 'parsed' is content with any number.

Quite a few tests were depending on minOccurs error detection though
occursCountKind was 'parsed'. Those test have been changed to use
occursCountKind='implicit'.

In addition, problems with grammar terms intended to absorb trailing
separators were found. These are not fixed. DFDL-110 is the ticket for
this
and will require substantial work as the current grammar productions
dont have information about "potentially trailing", so don't know when
to lay down things that can absorb separators or not, and also don't
know
about potential conflicts of separators from enclosing and enclosed
sequences.

DFDL-983