Relax restrictions on unbounded rep unparser
The RepUnboundedUnparser asserts that the current infoset node when
starting repetitions is a DIArray. However, this isn't always the case.
For example, if we have an array with minOccurs=1 and
maxOccurs=unbounded, then we will have a RepExactlyN unparser (with N =
1) followed by a RepUnbounded unparser. This means that when we get to
the RepUnbounded parser, the current infoset node will be one of the
elements in the array (i.e. DIComplex or DISimple) that was retrieved by
the RepExactlyN unparser. As such, the currentInfosetNode is potentially
meaningless in the RepUnboundedparser as far as asserting its in an
array. So remove these asserts.

This shouldn't affect anything, since the parent/child ElementKindParsers should
properly assert that the correct infoset elements/events are retrieved.

DFDL-1350